feat: isolating `chat-state` useEffect for retrigger in a hook
